// src/types/apps/chatTypes.ts
export type MsgFeedbackType = {
  isSent: boolean
  isSeen: boolean
  isDelivered: boolean
}
// ...
export type MessageType = {
  time: string | Date
  message: string
  senderId: number
  responseTime: string
  chatlogId: string
  history: any[]
  feedback: MsgFeedbackType
  question: string
}

export type ChatLogChatType = {
  msg: string
  time: string | Date
  responseTime: string
  chatlogId: string
  history: any[]
  feedback: MsgFeedbackType
  question: string
}

export type FormattedChatsType = {
  senderId: number
  messages: ChatLogChatType[]
}

export type MessageGroupType = {
  senderId: number
  messages: ChatLogChatType[]
}